+++ compiler/rustc_ast_lowering/src/stability.rs
@@ -3,7 +3,7 @@ use std::fmt;
 use rustc_abi::ExternAbi;
 use rustc_feature::Features;
 use rustc_session::Session;
-use rustc_session::parse::feature_err;
+use rustc_session::errors::feature_err;
 use rustc_span::symbol::sym;
 use rustc_span::{Span, Symbol};
