+++ compiler/rustc_session/src/parse.rs
@@ -1,7 +1,6 @@
 //! Contains `ParseSess` which holds state living beyond what one `Parser` might.
 //! It also serves as an input to the parser itself.
 
-use std::str;
 use std::sync::Arc;
 
 use rustc_ast::attr::AttrIdGenerator;
@@ -11,21 +10,15 @@ use rustc_data_structures::sync::{AppendOnlyVec, DynSend, DynSync, Lock};
 use rustc_errors::annotate_snippet_emitter_writer::AnnotateSnippetEmitter;
 use rustc_errors::emitter::{EmitterWithNote, stderr_destination};
 use rustc_errors::{
-    BufferedEarlyLint, ColorConfig, DecorateDiagCompat, Diag, DiagCtxt, DiagCtxtHandle,
-    DiagMessage, EmissionGuarantee, Level, MultiSpan, StashKey,
+    BufferedEarlyLint, ColorConfig, DecorateDiagCompat, Diag, DiagCtxt, DiagCtxtHandle, Level,
+    MultiSpan,
 };
-use rustc_feature::{GateIssue, find_feature_issue};
 use rustc_span::edition::Edition;
 use rustc_span::hygiene::ExpnId;
 use rustc_span::source_map::{FilePathMapping, SourceMap};
-use rustc_span::{Span, Symbol, sym};
+use rustc_span::{Span, Symbol};
 
 use crate::Session;
-use crate::errors::{
-    CliFeatureDiagnosticHelp, FeatureDiagnosticForIssue, FeatureDiagnosticHelp,
-    FeatureDiagnosticSuggestion, FeatureGateError, SuggestUpgradeCompiler,
-};
-use crate::lint::builtin::UNSTABLE_SYNTAX_PRE_EXPANSION;
 use crate::lint::{Lint, LintId};
 
 /// Collected spans during parsing for places where a certain feature was
@@ -76,169 +69,6 @@ impl SymbolGallery {
     pub fn insert(&self, symbol: Symbol, span: Span) {
         self.symbols.lock().entry(symbol).or_insert(span);
     }
-}
-
-// FIXME: this function now accepts `Session` instead of `ParseSess` and should be relocated
-/// Construct a diagnostic for a language feature error due to the given `span`.
-/// The `feature`'s `Symbol` is the one you used in `unstable.rs` and `rustc_span::symbol`.
-#[track_caller]
-pub fn feature_err(
-    sess: &Session,
-    feature: Symbol,
-    span: impl Into<MultiSpan>,
-    explain: impl Into<DiagMessage>,
-) -> Diag<'_> {
-    feature_err_issue(sess, feature, span, GateIssue::Language, explain)
-}
-
-/// Construct a diagnostic for a feature gate error.
-///
-/// This variant allows you to control whether it is a library or language feature.
-/// Almost always, you want to use this for a language feature. If so, prefer `feature_err`.
-#[track_caller]
-pub fn feature_err_issue(
-    sess: &Session,
-    feature: Symbol,
-    span: impl Into<MultiSpan>,
-    issue: GateIssue,
-    explain: impl Into<DiagMessage>,
-) -> Diag<'_> {
-    let span = span.into();
-
-    // Cancel an earlier warning for this same error, if it exists.
-    if let Some(span) = span.primary_span()
-        && let Some(err) = sess.dcx().steal_non_err(span, StashKey::EarlySyntaxWarning)
-    {
-        err.cancel()
-    }
-
-    let mut err = sess.dcx().create_err(FeatureGateError { span, explain: explain.into() });
-    add_feature_diagnostics_for_issue(&mut err, sess, feature, issue, false, None);
-    err
-}
-
-/// Construct a future incompatibility diagnostic for a feature gate.
-///
-/// This diagnostic is only a warning and *does not cause compilation to fail*.
-#[track_caller]
-pub fn feature_warn(sess: &Session, feature: Symbol, span: Span, explain: &'static str) {
-    feature_warn_issue(sess, feature, span, GateIssue::Language, explain);
-}
-
-/// Construct a future incompatibility diagnostic for a feature gate.
-///
-/// This diagnostic is only a warning and *does not cause compilation to fail*.
-///
-/// This variant allows you to control whether it is a library or language feature.
-/// Almost always, you want to use this for a language feature. If so, prefer `feature_warn`.
-#[track_caller]
-pub fn feature_warn_issue(
-    sess: &Session,
-    feature: Symbol,
-    span: Span,
-    issue: GateIssue,
-    explain: &'static str,
-) {
-    let mut err = sess.dcx().struct_span_warn(span, explain);
-    add_feature_diagnostics_for_issue(&mut err, sess, feature, issue, false, None);
-
-    // Decorate this as a future-incompatibility lint as in rustc_middle::lint::lint_level
-    let lint = UNSTABLE_SYNTAX_PRE_EXPANSION;
-    let future_incompatible = lint.future_incompatible.as_ref().unwrap();
-    err.is_lint(lint.name_lower(), /* has_future_breakage */ false);
-    err.warn(lint.desc);
-    err.note(format!("for more information, see {}", future_incompatible.reason.reference()));
-
-    // A later feature_err call can steal and cancel this warning.
-    err.stash(span, StashKey::EarlySyntaxWarning);
-}
-
-/// Adds the diagnostics for a feature to an existing error.
-/// Must be a language feature!
-pub fn add_feature_diagnostics<G: EmissionGuarantee>(
-    err: &mut Diag<'_, G>,
-    sess: &Session,
-    feature: Symbol,
-) {
-    add_feature_diagnostics_for_issue(err, sess, feature, GateIssue::Language, false, None);
-}
-
-/// Adds the diagnostics for a feature to an existing error.
-///
-/// This variant allows you to control whether it is a library or language feature.
-/// Almost always, you want to use this for a language feature. If so, prefer
-/// `add_feature_diagnostics`.
-pub fn add_feature_diagnostics_for_issue<G: EmissionGuarantee>(
-    err: &mut Diag<'_, G>,
-    sess: &Session,
-    feature: Symbol,
-    issue: GateIssue,
-    feature_from_cli: bool,
-    inject_span: Option<Span>,
-) {
-    if let Some(n) = find_feature_issue(feature, issue) {
-        err.subdiagnostic(FeatureDiagnosticForIssue { n });
-    }
-
-    // #23973: do not suggest `#![feature(...)]` if we are in beta/stable
-    if sess.unstable_features.is_nightly_build() {
-        if feature_from_cli {
-            err.subdiagnostic(CliFeatureDiagnosticHelp { feature });
-        } else if let Some(span) = inject_span {
-            err.subdiagnostic(FeatureDiagnosticSuggestion { feature, span });
-        } else {
-            err.subdiagnostic(FeatureDiagnosticHelp { feature });
-        }
-        if feature == sym::rustc_attrs {
-            // We're unlikely to stabilize something out of `rustc_attrs`
-            // without at least renaming it, so pointing out how old
-            // the compiler is will do little good.
-        } else if sess.opts.unstable_opts.ui_testing {
-            err.subdiagnostic(SuggestUpgradeCompiler::ui_testing());
-        } else if let Some(suggestion) = SuggestUpgradeCompiler::new() {
-            err.subdiagnostic(suggestion);
-        }
-    }
-}
-
-/// This is only used by unstable_feature_bound as it does not have issue number information for now.
-/// This is basically the same as `feature_err_issue`
-/// but without the feature issue note. If we can do a lookup for issue number from feature name,
-/// then we should directly use `feature_err_issue` for ambiguity error of
-/// `#[unstable_feature_bound]`.
-#[track_caller]
-pub fn feature_err_unstable_feature_bound(
-    sess: &Session,
-    feature: Symbol,
-    span: impl Into<MultiSpan>,
-    explain: impl Into<DiagMessage>,
-) -> Diag<'_> {
-    let span = span.into();
-
-    // Cancel an earlier warning for this same error, if it exists.
-    if let Some(span) = span.primary_span() {
-        if let Some(err) = sess.dcx().steal_non_err(span, StashKey::EarlySyntaxWarning) {
-            err.cancel()
-        }
-    }
-
-    let mut err = sess.dcx().create_err(FeatureGateError { span, explain: explain.into() });
-
-    // #23973: do not suggest `#![feature(...)]` if we are in beta/stable
-    if sess.unstable_features.is_nightly_build() {
-        err.subdiagnostic(FeatureDiagnosticHelp { feature });
-
-        if feature == sym::rustc_attrs {
-            // We're unlikely to stabilize something out of `rustc_attrs`
-            // without at least renaming it, so pointing out how old
-            // the compiler is will do little good.
-        } else if sess.opts.unstable_opts.ui_testing {
-            err.subdiagnostic(SuggestUpgradeCompiler::ui_testing());
-        } else if let Some(suggestion) = SuggestUpgradeCompiler::new() {
-            err.subdiagnostic(suggestion);
-        }
-    }
-    err
 }
 
 /// Info about a parsing session.
